+++ src/bitmap/bitmap.cpp
@@ -102,6 +102,7 @@ int main(int argc, char *argv[]) {
         {"mipmap-count", required_argument, 0, 'm' },
         {"mipmap-scale", required_argument, 0, 's' },
         {"mipmap-sharpen", required_argument, 0, 'P' },
+        {"mipmap-blur", required_argument, 0, 'U' },
         {"detail-fade", required_argument, 0, 'f' },
         {"budget", required_argument, 0, 'B' },
         {"budget-count", required_argument, 0, 'C' },
@@ -113,7 +114,7 @@ int main(int argc, char *argv[]) {
     };
 
     // Go through each argument
-    while((opt = getopt_long(argc, argv, "D:iIhd:t:f:s:f:F:m:T:S:B:C:p:h:u:H:P:", options, &longindex)) != -1) {
+    while((opt = getopt_long(argc, argv, "D:iIhd:t:f:s:f:F:m:T:S:B:C:p:h:u:H:P:U:", options, &longindex)) != -1) {
         switch(opt) {
             case 'd':
                 data = optarg;
@@ -270,6 +271,14 @@ int main(int argc, char *argv[]) {
                 }
                 break;
 
+            case 'U':
+                blur = static_cast<float>(std::strtof(optarg, nullptr));
+                if(blur.value() < 0.0F) {
+                    eprintf("Invalid blur value %f\n", blur.value());
+                    return EXIT_FAILURE;
+                }
+                break;
+
             case 'm':
                 max_mipmap_count = static_cast<std::int32_t>(std::strtol(optarg, nullptr, 10));
                 break;
@@ -309,7 +318,9 @@ int main(int argc, char *argv[]) {
                 eprintf("Mipmap options:\n");
                 eprintf("    --mipmap-count,-m <count>  Set maximum mipmaps. Default (new tag): 32767\n");
                 eprintf("    --mipmap-scale,-s <type>   Mipmap scale type. Can be: linear, nearest-alpha,\n");
-                eprintf("                               nearest. Default (new tag): linear\n\n");
+                eprintf("                               nearest. Default (new tag): linear\n");
+                eprintf("    --mipmap-blur,-U <radius>  Blur the bitmap before mipmapping. Higher radii\n");
+                eprintf("                               are exponentially slower. Default (new tag): 0.0\n");
                 eprintf("    --mipmap-sharpen,-P <amt>  Sharpen amount from 0 - 1. Default (new tag): 0.0\n");
                 eprintf("Bumpmap options (only applies to bumpmap bitmaps):\n");
                 eprintf("    --bump-height,-H <height>  Set the apparent bumpmap height from 0 to 1.\n");